Trait msgproc::prelude::config::Message

pub trait Message {
    type Headers: Headers;

    fn key(&self) -> Option<&[u8]>;
    fn payload(&self) -> Option<&[u8]>;
    unsafe fn payload_mut(&mut self) -> Option<&mut [u8]>;
    fn topic(&self) -> &str;
    fn partition(&self) -> i32;
    fn offset(&self) -> i64;
    fn timestamp(&self) -> Timestamp;
    fn headers(&self) -> Option<&Self::Headers>;

    fn payload_view<P>(&self) -> Option<Result<&P, <P as FromBytes>::Error>>
    where
        P: FromBytes + ?Sized,
    { ... }
    fn key_view<K>(&self) -> Option<Result<&K, <K as FromBytes>::Error>>
    where
        K: FromBytes + ?Sized,
    { ... }
}

A generic representation of a Kafka message.

Only read-only methods are provided by this trait, as the underlying storage might not allow modification.

Required Associated Types

type Headers: Headers

The type of headers that this message contains.

Required Methods

fn key(&self) -> Option<&[u8]>

Returns the key of the message, or None if there is no key.

fn payload(&self) -> Option<&[u8]>

Returns the payload of the message, or None if there is no payload.

unsafe fn payload_mut(&mut self) -> Option<&mut [u8]>

Returns a mutable reference to the payload of the message, or None if there is no payload.

Safety

librdkafka does not formally guarantee that modifying the payload is safe. Calling this method may therefore result in undefined behavior.

fn topic(&self) -> &str

Returns the source topic of the message.

fn partition(&self) -> i32

Returns the partition number where the message is stored.

fn offset(&self) -> i64

Returns the offset of the message within the partition.

fn timestamp(&self) -> Timestamp

Returns the message timestamp.

fn headers(&self) -> Option<&Self::Headers>

Returns the headers of the message, or None if there are no headers.

Provided Methods

fn payload_view<P>(&self) -> Option<Result<&P, <P as FromBytes>::Error>> where
P: FromBytes + ?Sized,

Converts the raw bytes of the payload to a reference of the specified type, that points to the same data inside the message and without performing any memory allocation.

fn key_view<K>(&self) -> Option<Result<&K, <K as FromBytes>::Error>> where
K: FromBytes + ?Sized,

Converts the raw bytes of the key to a reference of the specified type, that points to the same data inside the message and without performing any memory allocation.